Moving management related services to its own project
