+++ DotNET.Fundamentals.sln
@@ -177,6 +177,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DependencyInversion.Managem
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Grpc.CSharp", "Source\Grpc.CSharp\Grpc.CSharp.csproj", "{41A2B965-D8D9-420F-81E4-5E1C9F8D3CD6}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Management", "Source\Management\Management.csproj", "{CB2DBBE0-2FEA-43EC-BF47-3F6A2EF4B993}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -1222,6 +1224,18 @@ Global
 		{41A2B965-D8D9-420F-81E4-5E1C9F8D3CD6}.Release|x64.Build.0 = Release|Any CPU
 		{41A2B965-D8D9-420F-81E4-5E1C9F8D3CD6}.Release|x86.ActiveCfg = Release|Any CPU
 		{41A2B965-D8D9-420F-81E4-5E1C9F8D3CD6}.Release|x86.Build.0 = Release|Any CPU
+		{CB2DBBE0-2FEA-43EC-BF47-3F6A2EF4B993}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{CB2DBBE0-2FEA-43EC-BF47-3F6A2EF4B993}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{CB2DBBE0-2FEA-43EC-BF47-3F6A2EF4B993}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{CB2DBBE0-2FEA-43EC-BF47-3F6A2EF4B993}.Debug|x64.Build.0 = Debug|Any CPU
+		{CB2DBBE0-2FEA-43EC-BF47-3F6A2EF4B993}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{CB2DBBE0-2FEA-43EC-BF47-3F6A2EF4B993}.Debug|x86.Build.0 = Debug|Any CPU
+		{CB2DBBE0-2FEA-43EC-BF47-3F6A2EF4B993}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{CB2DBBE0-2FEA-43EC-BF47-3F6A2EF4B993}.Release|Any CPU.Build.0 = Release|Any CPU
+		{CB2DBBE0-2FEA-43EC-BF47-3F6A2EF4B993}.Release|x64.ActiveCfg = Release|Any CPU
+		{CB2DBBE0-2FEA-43EC-BF47-3F6A2EF4B993}.Release|x64.Build.0 = Release|Any CPU
+		{CB2DBBE0-2FEA-43EC-BF47-3F6A2EF4B993}.Release|x86.ActiveCfg = Release|Any CPU
+		{CB2DBBE0-2FEA-43EC-BF47-3F6A2EF4B993}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{AAA185BC-0EA0-4804-9D7B-0E9B1D411070} = {780A06A2-8A9B-49DA-AA18-472C4F712673}
@@ -1309,5 +1323,6 @@ Global
 		{8240278D-3B83-487D-9760-24675DC81EAA} = {780A06A2-8A9B-49DA-AA18-472C4F712673}
 		{E288C52F-A293-4ECC-B5DA-6825261F8434} = {780A06A2-8A9B-49DA-AA18-472C4F712673}
 		{41A2B965-D8D9-420F-81E4-5E1C9F8D3CD6} = {780A06A2-8A9B-49DA-AA18-472C4F712673}
+		{CB2DBBE0-2FEA-43EC-BF47-3F6A2EF4B993} = {780A06A2-8A9B-49DA-AA18-472C4F712673}
 	EndGlobalSection
 EndGlobal
