+++ Source/Management/ICanBindManagementServices.cs
@@ -2,7 +2,9 @@
  *  Copyright (c) Dolittle. All rights reserved.
  *  Licensed under the MIT License. See LICENSE in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
-namespace Dolittle.Hosting
+using Dolittle.Hosting;
+
+namespace Dolittle.Management
 {
     /// <summary>
     /// Defines a system that can bind a gRPC service for management purpose
