+++ Source/Management/Management.csproj
@@ -1,0 +1,13 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>netstandard2.0</TargetFramework>
+    <AssemblyName>Dolittle.Management</AssemblyName>
+  </PropertyGroup>
+
+  <Import Project="../../Build/MSBuild/default.props"/>
+
+  <ItemGroup>
+    <ProjectReference Include="../Hosting/Hosting.csproj"/>
+  </ItemGroup>
+</Project>
